requirements.txt
openai == 1.55.3
python-dotenv == 1.0.1
python-telegram-bot==21.7
nest_asyncio == 1.6.0
mysql-connector-python==9.1.0
ymaps==1.3
aiohttp
requests
